# v3/api/openapi.yaml
openapi: 3.1.0
info:
  title: Sakura v3 API
  version: 3.0.0-alpha.1
  description: Unified API used equally by the Web portal, Telegram Bot and API clients.
servers:
  - url: /
paths:
...
  /api/v3/me:
    get:
      operationId: getCurrentAccount
      tags: [Accounts]
      security: [{cookieAuth: []}]
      responses:
        "200": {description: Unified account and identities}
...
  /api/v3/me/tickets:
    get:
      operationId: listMyTickets
      tags: [Tickets]
      security: [{cookieAuth: []}]
      responses: {"200": {description: Tickets owned by the current account}}
    post:
      operationId: createTicket
      tags: [Tickets]
      security: [{cookieAuth: [], csrfHeader: []}]
      responses: {"201": {description: Ticket and first public message created}}
  /api/v3/me/tickets/{id}:
    parameters: [{$ref: "#/components/parameters/ID"}]
    get:
      operationId: getMyTicket
      tags: [Tickets]
      security: [{cookieAuth: []}]
      responses: {"200": {description: Ticket owned by the current account}}
  /api/v3/me/tickets/{id}/messages:
    parameters: [{$ref: "#/components/parameters/ID"}]
    get:
      operationId: listMyTicketMessages
      tags: [Tickets]
      security: [{cookieAuth: []}]
      responses: {"200": {description: Public messages only; internal staff notes are always excluded}}
    post:
      operationId: replyToMyTicket
      tags: [Tickets]
      security: [{cookieAuth: [], csrfHeader: []}]
      responses: {"201": {description: Public user reply created}}
  /api/v3/admin/tickets:
    get:
      operationId: listTickets
      tags: [Administration, Tickets]
      security: [{cookieAuth: []}]
      responses: {"200": {description: Ticket operations queue}}
  /api/v3/admin/tickets/{id}:
    parameters: [{$ref: "#/components/parameters/ID"}]
    get:
      operationId: getTicket
      tags: [Administration, Tickets]
      security: [{cookieAuth: []}]
      responses: {"200": {description: Ticket detail}}
    put:
      operationId: updateTicket
      tags: [Administration, Tickets]
      security: [{cookieAuth: [], csrfHeader: []}]
      responses: {"200": {description: Ticket state, priority and assignment updated}}
  /api/v3/admin/tickets/{id}/messages:
    parameters: [{$ref: "#/components/parameters/ID"}]
    get:
      operationId: listTicketMessages
      tags: [Administration, Tickets]
      security: [{cookieAuth: []}]
      responses: {"200": {description: Public replies and internal notes}}
    post:
      operationId: addTicketMessage
      tags: [Administration, Tickets]
      security: [{cookieAuth: [], csrfHeader: []}]
      responses: {"201": {description: Public reply or internal staff note created}}
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sakura_v3_session
    csrfHeader:
      type: apiKey
      in: header
      name: X-CSRF-Token
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
...
    ID:
      name: id
      in: path
      required: true
      schema: {type: string, format: uuid}
...